+++ src/InflectorConfigServiceProvider.php
@@ -6,7 +6,8 @@ use League\Container\ServiceProvider\AbstractServiceProvider;
 use League\Container\ServiceProvider\BootableServiceProviderInterface;
 
 class InflectorConfigServiceProvider extends AbstractServiceProvider implements
-    BootableServiceProviderInterface
+    BootableServiceProviderInterface,
+    ConfigurableServiceProvider
 {
     /**
      * @var array
@@ -19,6 +20,14 @@ class InflectorConfigServiceProvider extends AbstractServiceProvider implements
      * @param array  $config
      */
     public function __construct(array $config)
+    {
+        $this->config = $config;
+    }
+
+    /**
+     * @param array  $config
+     */
+    public function configure(array $config)
     {
         $this->config = $config;
     }