+++ tests/ConfigServiceProviderTest.php
@@ -5,6 +5,7 @@ namespace tests\TomPHP\ConfigServiceProvider;
 use League\Container\Container;
 use PHPUnit_Framework_TestCase;
 use TomPHP\ConfigServiceProvider\ConfigServiceProvider;
+use Prophecy\Argument;
 
 class ConfigServiceProviderTest extends PHPUnit_Framework_TestCase
 {
@@ -16,6 +17,13 @@ class ConfigServiceProviderTest extends PHPUnit_Framework_TestCase
     protected function setUp()
     {
         $this->container = new Container();
+
+        $this->subProvider = $this->prophesize('TomPHP\ConfigServiceProvider\ConfigurableServiceProvider');
+
+        $this->subProvider->configure(Argument::any())->willReturn();
+        $this->subProvider->provides()->willReturn([]);
+        $this->subProvider->setContainer(Argument::any())->willReturn();
+        $this->subProvider->register()->willReturn();
     }
 
     public function testItProvidesConfigValuesViaTheDI()
@@ -109,6 +117,150 @@ class ConfigServiceProviderTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(
             'test value',
             $this->container->get('config/test_group/test_setting')
+        );
+    }
+
+    /**
+     * @group sub_providers
+     */
+    public function testItConfiguresASubProvider()
+    {
+        $config = [
+            'sub_provider' => ['key' => 'config'],
+        ];
+
+        new ConfigServiceProvider($config, 'config', '.', [
+            'sub_provider' => $this->subProvider->reveal(),
+        ]);
+
+        $this->subProvider->configure(['key' => 'config'])->shouldHaveBeenCalled();
+    }
+
+    /**
+     * @group sub_providers
+     */
+    public function testItSkipsConfiguringASubProvderWithNoConfig()
+    {
+        new ConfigServiceProvider([], 'config', '.', [
+            'sub_provider' => $this->subProvider->reveal(),
+        ]);
+
+        $this->subProvider->configure(Argument::any())->shouldNotHaveBeenCalled();
+    }
+
+    /**
+     * @group sub_providers
+     */
+    public function testItMergesTheSubProvidersServiceList()
+    {
+        $this->subProvider->provides()->willReturn(['b']);
+
+        $provider = new ConfigServiceProvider(
+            ['sub_provider' => [], 'a' => 1],
+            'config',
+            '.',
+            [ 'sub_provider' => $this->subProvider->reveal() ]
+        );
+
+        $this->assertEquals(['config.sub_provider', 'config.a', 'b'], $provider->provides());
+    }
+
+    /**
+     * @group sub_providers
+     */
+    public function testItRegistersSubProviders()
+    {
+        $this->container->addServiceProvider(new ConfigServiceProvider(
+            ['sub_provider' => []],
+            'config',
+            '.',
+            [ 'sub_provider' => $this->subProvider->reveal() ]
+        ));
+
+        $this->container->get('config.sub_provider');
+
+        $this->subProvider->setContainer($this->container)->shouldHaveBeenCalled();
+        $this->subProvider->register()->shouldHaveBeenCalled();
+    }
+
+    /**
+     * @group sub_providers
+     */
+    public function testBootableSubProvidersAreBooted()
+    {
+        $this->subProvider = $this->prophesize('tests\mocks\BootableConfigurableServiceProvider');
+
+        $this->subProvider->configure(Argument::any())->willReturn();
+        $this->subProvider->provides()->willReturn([]);
+        $this->subProvider->setContainer(Argument::any())->willReturn();
+        $this->subProvider->register()->willReturn();
+        $this->subProvider->boot()->willReturn();
+
+        $this->container->addServiceProvider(new ConfigServiceProvider(
+            ['sub_provider' => []],
+            'config',
+            '.',
+            [ 'sub_provider' => $this->subProvider->reveal() ]
+        ));
+
+        $this->container->get('config.sub_provider');
+
+        $this->subProvider->setContainer($this->container)->shouldHaveBeenCalled();
+        $this->subProvider->boot()->shouldHaveBeenCalled();
+    }
+
+    /**
+     * @group from_config_factory
+     */
+    public function testItCreatesFromConfigWithDefaultSettings()
+    {
+        $config = [
+            'test_key' => 'test value',
+
+            'inflectors' => [
+                'tests\mocks\ExampleInterface' => [
+                    'setValue' => ['config.test_key']
+                ]
+            ]
+        ];
+
+        $this->container->addServiceProvider(ConfigServiceProvider::fromConfig($config));
+
+        $this->container->add('example', 'tests\mocks\ExampleClass');
+
+        $this->assertEquals(
+            'test value',
+            $this->container->get('example')->getValue()
+        );
+    }
+
+    /**
+     * @group from_config_factory
+     */
+    public function testItCanOverrideDefaultSettings()
+    {
+        $config = [
+            'test_key' => 'test value',
+
+            'inflectors' => [
+                'tests\mocks\ExampleInterface' => [
+                    'setValue' => ['settings/test_key']
+                ]
+            ]
+        ];
+
+        $this->container->addServiceProvider(
+            ConfigServiceProvider::fromConfig($config, [
+                'prefix' => 'settings',
+                'separator' => '/'
+            ])
+        );
+
+        $this->container->add('example', 'tests\mocks\ExampleClass');
+
+        $this->assertEquals(
+            'test value',
+            $this->container->get('example')->getValue()
         );
     }
 }