+++ tests/InflectorConfigServiceProviderTest.php
@@ -18,6 +18,14 @@ class InflectorConfigServiceProviderTest extends PHPUnit_Framework_TestCase
         $this->container = new Container();
     }
 
+    public function testItIsAConfigurableServiceProvider()
+    {
+        $this->assertInstanceOf(
+            'TomPHP\ConfigServiceProvider\ConfigurableServiceProvider',
+            new InflectorConfigServiceProvider([])
+        );
+    }
+
     public function testSetsUpAnInflector()
     {
         $config = [
@@ -27,6 +35,26 @@ class InflectorConfigServiceProviderTest extends PHPUnit_Framework_TestCase
         ];
 
         $this->container->addServiceProvider(new InflectorConfigServiceProvider($config));
+        $this->container->add('example', 'tests\mocks\ExampleClass');
+
+        $this->assertEquals(
+            'test_value',
+            $this->container->get('example')->getValue()
+        );
+    }
+
+    public function testCanBeReconfigured()
+    {
+        $config = [
+            'tests\mocks\ExampleInterface' => [
+                'setValue' => ['test_value']
+            ]
+        ];
+
+        $provider = new InflectorConfigServiceProvider([]);
+        $provider->configure($config);
+
+        $this->container->addServiceProvider($provider);
         $this->container->add('example', 'tests\mocks\ExampleClass');
 
         $this->assertEquals(