+++ tests/mocks/BootableConfigurableServiceProvider.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace tests\mocks;
+
+use League\Container\ServiceProvider\BootableServiceProviderInterface;
+use TomPHP\ConfigServiceProvider\ConfigurableServiceProvider;
+
+interface BootableConfigurableServiceProvider extends
+    BootableServiceProviderInterface,
+    ConfigurableServiceProvider
+{
+}